move serialize to compiler
